4. Реализовал слой сервиса

+++ src/main/java/ru/javawebinar/topjava/model/Meal.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 public class Meal extends AbstractBaseEntity {
-    private final long userId;
+    private final int userId;
 
     private final LocalDateTime dateTime;
 
@@ -13,11 +13,11 @@ public class Meal extends AbstractBaseEntity {
 
     private final int calories;
 
-    public Meal(LocalDateTime dateTime, String description, int calories, long userId) {
+    public Meal(LocalDateTime dateTime, String description, int calories, int userId) {
         this(null, dateTime, description, calories, userId);
     }
 
-    public Meal(Integer id, LocalDateTime dateTime, String description, int calories, long userId) {
+    public Meal(Integer id, LocalDateTime dateTime, String description, int calories, int userId) {
         super(id);
         this.id = id;
         this.dateTime = dateTime;
